refactor: Remove the use of interfaces
