+++ utils/component/generator.go
@@ -1,10 +1,6 @@
-package compgen
+package component
 
 import (
-	// "fmt"
-
-	"fmt"
-
 	"github.com/layer5io/meshkit/models/meshmodel/core/v1alpha1"
 	"github.com/layer5io/meshkit/utils"
 	"github.com/sirupsen/logrus"
@@ -22,18 +18,16 @@ type CuePathConfig struct {
 	IdentifierPath string
 }
 
-type prometheusComponentsGenerator struct {
+type ComponentsGenerator struct {
 	crds                []string
 	extractorPathConfig CuePathConfig
 	resourceMetadata    map[string]interface{}
 }
 
-func (pcg prometheusComponentsGenerator) Generate() ([]v1alpha1.Component, error) {
+func (cg ComponentsGenerator) Generate() ([]v1alpha1.Component, error) {
 	components := make([]v1alpha1.Component, 0)
-	for _, crd := range pcg.crds {
-		meta := map[string]interface{}{
-			"icon": DefaultPrometheusComponentIconURL,
-		}
+	for _, crd := range cg.crds {
+		meta := cg.resourceMetadata
 		comp := v1alpha1.NewComponent()
 		// crds should be yaml
 		crdCue, err := utils.YamlToCue(crd)
@@ -42,28 +36,28 @@ func (pcg prometheusComponentsGenerator) Generate() ([]v1alpha1.Component, error
 			continue
 		}
 
-		schema, err := getSchema(crdCue, pcg.extractorPathConfig)
+		schema, err := getSchema(crdCue, cg.extractorPathConfig)
 		if err != nil {
 			logrus.Warn(ErrCrdGenerate(err))
 			continue
 		}
 		comp.Spec = schema
-		name, err := extractValueFromPath(crdCue, pcg.extractorPathConfig.NamePath)
-		fmt.Println("name", name)
+		name, err := extractValueFromPath(crdCue, cg.extractorPathConfig.NamePath)
 		if err != nil {
 			logrus.Warn(ErrCrdGenerate(err))
 			continue
 		}
 		meta["name"] = name
 		// append given metadata
-		for k, v := range pcg.resourceMetadata {
+		for k, v := range cg.resourceMetadata {
 			meta[k] = v
 		}
 		comp.Metadata = meta
+		components = append(components, comp)
 	}
 	return components, nil
 }
 
-func NewPrometheusComponentsGenerator(crds []string, pathConf CuePathConfig) ComponentsGenerator {
-	return &prometheusComponentsGenerator{crds: crds, extractorPathConfig: pathConf}
+func NewComponentsGenerator(crds []string, pathConf CuePathConfig, meta map[string]interface{}) *ComponentsGenerator {
+	return &ComponentsGenerator{crds: crds, extractorPathConfig: pathConf, resourceMetadata: meta}
 }
